previous/V2-convbase/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.9.2
pandas>=1.3.0
Pillow>=8.3.1
tqdm>=4.62.0
albumentations>=1.3.1
numpy>=1.21.0
scikit-learn>=1.0.0
